# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Contacts API",
    "description": "Contacts project backend service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://contacts-29l4.onrender.com",
      "description": "production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "local server"
    }
  ],
...
  "paths": {
...
    "/api/users/": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "subscription": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/contacts/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "favorite",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "favorite": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}